requirements/base.txt
# Non-Django requirements
ipython==2.1.0

# Django requirements
git+https://github.com/django/django.git@1.7c2
git+https://github.com/feincms/feincms.git@next
feincms-elephantblog>=1.0.0
django-compressor==1.4
git+https://github.com/ebrelsford/django-inplace.git@master
django-moderation==0.3.6
djangorestframework==2.3.14
git+https://github.com/djangonauts/django-rest-framework-gis.git@master
